remove the corridor ensonification code
Before this commit, the code for the corridor ensonification lived in this R12 folder. Initially, I had organized the code this way to make it easier for Hadley to work with the code. However, now I will take up the dragonfly experiments again and I want to move code not belonging to the R12 library to a different repo.